Add bookmarking and bookmarks view
Based heavily on the work for favorites.

Adds bookmarking functionality to the application. The status view has been updated to include a bookmark button that can bookmark/unbookmark a status. The profile page has been updated to include a button in the header to navigate to a page that lists your bookmarks.